CR.Elastos-SmartWeb-Service/requirements.txt
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
requests==2.22.0
psycopg2==2.8.2